# niki/target/niki/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "niki"
    },
    "host": "",
    "basePath": "/niki/rest",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/positions": {
            "get": {
                "consumes": [
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "max",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "start",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Position"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Position"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "headers": {
                            "Location": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/positions/{id}": {
            "get": {
                "consumes": [
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "$ref": "#/definitions/Position"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "headers": {
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Position"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "headers": {
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "headers": {
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "headers": {
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "headers": {
                        },
                        "schema": {
                            "$ref": "#/definitions/Position"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                ],
                "produces": [
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "headers": {
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "headers": {
                        }
                    }
                }
            }
        },
        "/provinces": {
            "get": {
                "consumes": [
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "max",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "start",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Province"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Province"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "headers": {
                            "Location": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/provinces/{id}": {
            "get": {
                "consumes": [
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                        },
                        "schema": {
                            "$ref": "#/definitions/Province"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "headers": {
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Province"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "headers": {
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "headers": {
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "headers": {
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "headers": {
                        },
                        "schema": {
                            "$ref": "#/definitions/Province"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                ],
                "produces": [
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "headers": {
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "headers": {
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Country": {
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "Position": {
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "Province": {
            "properties": {
                "code": {
                    "type": "string"
                },
                "country": {
                    "$ref": "#/definitions/Country"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
    }
}